Update to v3 endpoint to fetch data
This change also moves signal processing code from the webhook module
into the coordinator so it's shared between the two update paths.
